Revert "(CR): msm/mdss: HBM: fix 64 bit compilation issue"
This reverts commit b08adb9ff7176f8ed6edb5e8d3ab59c81a6b7d78.

Reviewed-on: http://gerrit.mot.com/840573
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Tested-by: Jira Key <jirakey@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>

+++ drivers/video/msm/mdss/mdss_mdp_overlay.c
@@ -2501,7 +2501,7 @@ static ssize_t hbm_store(struct device *dev,
 
 	mdss_mdp_clk_ctrl(MDP_BLOCK_POWER_ON);
 	r = mdss_mdp_ctl_intf_event(ctl, MDSS_EVENT_ENABLE_HBM,
-				(void *)(unsigned long)enable);
+				(void *)enable);
 	mdss_mdp_clk_ctrl(MDP_BLOCK_POWER_OFF);
 	if (r) {
 		pr_err("Failed sending HBM command, r = %d\n", r);
